blk/aio: allocate iocb using small_vector
there is chance that the batch size is so large that we have stack
overflow when allocating piocb array using VLA. in this change,
piocb is allocated using small_vector<>, as in most circumstances, the
size of an object is 4M when it comes to CephFS, RGW and RBD by default.
so if the client sends 4K / 8K aligned read / write, in general,
that would be translated into 1024 IO requests at most. so,
use 1024 as the size of small_vector would suffice.

also, 1024 is the default bdev_aio_max_queue_depth by default, hence
this setting is able to feed a io_submit() batch.

Fixes: https://tracker.ceph.com/issues/46366
Signed-off-by: Kefu Chai <kchai@redhat.com>